Made changes to please the Hound.
